+++ controlplane/test/subgraph/publish-subgraph.test.ts
@@ -100,6 +100,7 @@ describe('Publish subgraph tests', () => {
     'namespace-viewer',
     'graph-admin',
     'graph-viewer',
+    'subgraph-viewer',
   ])('%s should not be able to publish to existing regular subgraph', async (role) => {
     const { client, server, authenticator, users } = await SetupTest({ dbname });
 
@@ -236,6 +237,7 @@ describe('Publish subgraph tests', () => {
     'graph-admin',
     'graph-viewer',
     'subgraph-publisher',
+    'subgraph-viewer',
   ])('%s should not be able to publish regular subgraph without already being created', async (role) => {
     const { client, server, authenticator, users } = await SetupTest({ dbname });