+++ controlplane/CHANGELOG.md
@@ -4,6 +4,18 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.153.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.152.0...controlplane@0.153.0) (2025-07-10)
+
+### Features
+
+* add the `subgraph-viewer` role ([#2022](https://github.com/wundergraph/cosmo/issues/2022)) ([902b3bc](https://github.com/wundergraph/cosmo/commit/902b3bcd8ae9efaab71f5e1a97012938af51a57f)) (@wilsonrivera)
+
+# [0.152.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.151.2...controlplane@0.152.0) (2025-07-09)
+
+### Features
+
+* return the proposal name on the creation of proposal ([#2003](https://github.com/wundergraph/cosmo/issues/2003)) ([ff1b237](https://github.com/wundergraph/cosmo/commit/ff1b2376921db2db7ac2a9d1619824eb1ae1e76d)) (@JivusAyrus)
+
 ## [0.151.2](https://github.com/wundergraph/cosmo/compare/controlplane@0.151.1...controlplane@0.151.2) (2025-07-03)
 
 **Note:** Version bump only for package controlplane